Extract createProjectLifecycleSlice from projectStore (P5 store slice)
